Abbreviate module types when printing error messages (#1895)
* Abbreviate module types when printing error messages

Module types can become huge and pretty-printing them in error messages
seems to be very slow. In one case, we produced a ~400k lines error
message which took about 4 minutes to generate.

This PR caps the overall number of signature elements we print and the
depth to which we unfold module types. In my somewhat limited testing,
this seemed to be the best heuristic I tried, although it is far from
perfect. We can adjust later based on user feedback.

* Add comments

---------

Co-authored-by: Roman Leshchinskiy <rleshchinskiy@janestreet.com>