Penalise ambiguous paths when computing their size
This unifies the two forms of badness that we want to avoid when printing paths.
In particular, it allows searching through longer paths in hopes of avoiding an
ambiguous path, just as we prefer a longer path to a double-underscore path.
Fixes a test where the more agressive pruning in this PR caused us to settle on
an ambiguous path.