Partially fix Closure
This makes the code output for Closure not complete gibberish. Unfortunately,
it's fatally flawed, as it causes the .cmx for an argument module to be full of
zeros (which will immediately be mutated to their actual values, but Closure
isn't aware of this).

This probably isn't worth actually fixing, since (a) we don't plan to keep
Closure around much longer in ocaml-jst, and (b) this whole double-module-block
design probably needs to be revisited.