Better heuristic for resimplification after mutable unboxing
The new heuristic for triggering a resimplification after mutable
unboxing ensures that it is triggered whenever an actually mutable block
has been unboxed, even if not primitive was removed. This is because
when we unbox a mutable block that contains for instance a non-mutable
field, then we can propagate new information about that field, and it
warrants a new traversal.