Use unnamed field instead of double module block
Previously, an argument module would have a special format for its module block,
with exactly two fields: the module block that would normally be there, and a
second module block whose type matches the parameter exactly. Effectively, this
commit unboxes the first field, and also instead of storing the argument form of
the module directly, it stores the coercion from the module to its parameter
type as a function. The downside of this is that there won't be a fixed location
for the argument form of the module, since it will be recomputed each time the
module is used as an argument, but in practice this shouldn't matter often since
the computation will happen at compile time and the only circumstance in which
sharing would matter would be if the parameter is used as a first-class module.

Storing the argument form directly in this way could be tricky in cases where
the coercion is trivial, since then the module block will have a circular
reference to itself as one of its own fields. We could solve this by, rather
than passing the module block itself, rebuilding a copy of the module block and
passing that to the coercion.