Store argument block itself rather than coercion function
This is simpler and guarantees we never waste space by constructing the argument
block more than once. It's still a bit wasteful if the coercion is
trivial (i.e., if the module's type has all the same values as the parameter it
implements) but this seems unlikely to matter in practice.