Remove code gen changes from `-as-argument-for` branch
None of the changes to code gen can actually be tested without
`-instantiate`, which won't be implemented until PR #1905. This lets PR #2177
just be about typechecking, which we _can_ test immediately (in that we can
check that if you compile with `-as-argument-for` then you have to conform to
the parameter type and we remember that you passed the flag).