Move code between Printast and Pprintast (#2194)
* Revert "Pprintast now depends on Printast"

This reverts commit bdac10c4886aa4b4ff857274aa63f7bb86ef98e0.

That commit was part of #1417, but ended up getting squashed
in a rebase. Why revert? Because it's hard to upstream this
change, due to the new dependency of the tyvar printer on the
lexer. This can be disentangled, but doing so would be disruptive
for little gain. So I'm just undoing this unforced little cleanup,
which isn't so important anyway.

* Fix the build