Describe runtime parameters in .cmx/.cmo
This adds to the top-level `program` record in Lambda a description of the
runtime parameters a parameterised module takes. In particular, rather than just
giving the names of the parameters, specify whether each one should be satisfied
by an argument module or a dependency. This simplifies instantiation since there
is a difference in how the argument is calculated (for an argument module, we
need to project out the argument block). It also allows us to pass a unit value
when instantiating a module that doesn't actually use any of its parameters.