Fix confusingly-named function `depth_may_be_at_least`
The function actually computes a (possible) _strict_ inequality, which is crucial
for inlining heuristics: since the depth of a non-recursive function is zero and
the default max rec depth is zero, if we were allowing equality here, we would
inline nothing automatically.

Renamed to `depth_may_exceed`, which makes the calling code obviously correct.

Also changed `>= 1` to `> 0` to be more idiomatic (so that it also _looks_ like
it's computing a strict inequality).