Fix classification of `interp.c` in upstream Makefile
Previously, `interp.c` was classified as both a C source common to
native/bytecode AND a C source that was only needed in bytecode.
This is clearly confusing and wrong. Upstream just classifies it
as a bytecode-only C source. The build passes if I change that.
I'll check with Mark before I merge this.