flambda-backend: Represent mixed records as "faux mixed blocks" in bytecode (#2476)
* Runtime 4

* Runtime 5

* A little less weird behavior in runtime 4

* Move the documentation of 'faux' to a better place, and minimize diff

* Fix bug where dummy argument wasn't being pushed

* make boot

* Fix classification of `interp.c` in upstream Makefile

Previously, `interp.c` was classified as both a C source common to
native/bytecode AND a C source that was only needed in bytecode.
This is clearly confusing and wrong. Upstream just classifies it
as a bytecode-only C source. The build passes if I change that.
I'll check with Mark before I merge this.