Don't register the argument type as a parameter import
Registering the argument to `-as-argument-for` as a parameter import was
something of a relic of an era before `Persistent_env` had nicely separated
the reading of a signature from a .cmi from the binding of that .cmi's module in
the environment. Nowadays we can defer checking whether a loaded .cmi is a
parameter until it's referred to in source, so that if one is reading a .cmi for
any other purpose, there's no reason to bother with that registration step.

So this patch moves the awkward check in `Persistent_env` (putting one aspect of
it into `Typemod` where it belonged anyway), clarifies the API around
imported parameters, and stops passing the `-as-argument-for` module to
`Env.register_parameter_import`.