Fix missing rewrite ids when unboxing detect invalid apply conts
In some cases (e.g. a do_not_unbox decision, followed by a decision that
does unbox, and that discovers that some apply conts are invalids), it
could happen that the set of rewrite ids known as being invalid was
dropped/reset to empty. That together with the caching of the extra args
computsion (done through the rewrite ids seen) meant that we could
"forget" some rewrite ids.