Allow overlap of invalid rewrite id and extra_args
As the comment states, it can happen that, when adding a extra param and
args, there is an overlap between the domain of the extra_args map, and
the set of invalids (because of the way these are computed by the
unboxing code). This case is reasonable, so in such a case, we can allow
the invalid set to take precedenhce.