Don't infer tail-calls for applications where the called function contains a dot (#2781)
* Don't infer tail calls for applications with `.` in the function name

Suppose a function `f` allocates locals, and `f` calls `g` in tail
position. If `g` is inferred to be a tail-call, it cannot accept (as
parameters) locals that `f` had allocated as `f`'s region does not fully
enclose `g`'s region.

There are many places where function calls happen to be in tail
position, but do not need to be tail calls. These function calls
currently cannot accept locals due to the above restriction. That
restriction is annoying. E.g. the following function does not compile
(before this change), but we would like it to:

```ocaml
module rec List : sig
  val map : local_ ('a -> 'b) -> 'a list -> 'b list
end = List

let call_local_map_in_tail_pos lst =
  let local_ str = "Hello, world!" in
  let local_ f x = let _ = str in () in
  List.map f lst
```

Error:

```
Line 8, characters 11-12:
8 |   List.map f lst
               ^
Error: This value escapes its region.
  Hint: This argument cannot be local,
  because it is an argument in a tail call.
```

By inferring fewer tail calls, we can make that restriction less
annoying. Our first change is to (heuristically) not infer tail calls
for applications of functions that contain a dot (`.`) in the
identifier, as they are unlikely to require tail call optimization. We
are in the process of developing a better heuristic.

* Address code review comments

* Fix failing tests

* frome -> frame

* Print out dlambda instead of desl

* Simplify basic test stanza

* Update comparison_table.heap.reference

* Delete basic.reference