Reorder fields in block header
Place wosize above scannable prefix, simplifying expressions like
[Array.length].

[Array.length] and other block-size-computing expressions are fairly common in
real code, and this shrinks them in code size and also operations executed, a
rough diff being before:

      movabs  rbx, 72057594037927935
        mov     rax, qword ptr [rax - 8]
        and     rax, rbx
        shr     rax, 9
        or      rax, 1

and after:

	mov    rdx,QWORD PTR [rax-0x8]
 	shr    rdx,0x11
 	or     rdx,0x1

In addition I think this is just more readable when trying to pore over
generated assembly.

Getting the scannable prefix size is, I think, as bad as getting block sizes is
now: but the only time we do that is in GC scanning, and we also need lengths,
so the total cost should be unch'd.

I tested both runtime4 and runtime5 via `make runtest`, and inspected some
amount of generated assembly.