Fix for allocation of closures containing unboxed values (#2997)
* Fix for allocation of closures containings unboxed values

The current code for allocating dynamic sets of closure did not properly
take into account the fact that some of the env fields can now be
unbxoed values. In some cases (particularly when the sets are bigger
than Max_young_wosize), these unboxed fields require cmm code that sets
the corresponding field, and that store operation need to be annotated
with the correct memory_chunk.

This commit adds the necessary code to propagate enough information so
that we have access to the memory_chunks of every field when allocating
a dynmaic set of closures (static closures do not need that, which is
helpful since it would be much harder to produce the memory_chunks in
those cases).

At the same time, it reworks the current way these allocations are
generated: the old code used a `set` function, which in effect was only
there to distinguish on what is essentially the memory_chunk of the
value being written and use the appropriate cmm expression to write the
value. The code is simplified by instead providing the adequate
memory_chunk, from which the set function can be deduced.

* Take into account the actual size of env slots in closures

* Remove option type for chunks in to_cmm_set_of_closures

* fixes

* Rename the type/cstrs for regular blocks vs mixed blocks

* Rename value slot in slot_offsets

Cherry_picked from PR#2983

* typo

* typos

* Add doc + assertion for the shape of a mixed block

* Fix computation of subkinds for mixed blocks fields

* Code review

---------

Co-authored-by: Mark Shinwell <mshinwell@pm.me>