Abstracting over datatypes with higher kinds (#2889)
A factoring is when a higher kind is converted in and out of a parameter representation.

Variances don't work. They are ignored entirely if a factoring happened.

No syntax for datatype, need to use an attribute.

You cannot abstract a higher kind with a [@@datatype], as parameters don't get factored out properly.