Fix test with backtrace that embeds 17 line numbers from the compiler
The expected output for `tool-toplevel/pr6468.ml` has a backtrace with a bunch
of frames deep in the bowels of the compiler, wherever someone happened to raise
`Not_found` most recently. This is of course always a hazard with `Not_found`.
We can avoid it by just declaring and raising our own generic exception instead.
(There's still a line in `topeval.ml` but at least that doesn't change nearly as
often.)