Do not write to ring after going out of stw participant set. (#13529)
When the runtime events ring is destroyed, only the domains that are
part of the stop-the-world (stw) participant list are stopped. So it is
unsafe to write to ring when the domain is not part of the stw
participant set.

This fixes the data race in `lib-runtime-events/test_dropped_events`.

(cherry picked from commit 27c82c430ac27873e073ff7eab93a812d8fe67c8)