Move `backend/asm*` into `asmcomp/` (#3149)
* Move `backend/asm*` into `asmcomp/`

We're preparing to move upstream code out of `ocaml/` and realign this repo with
the upstream directory structure to make dealing with merges and history easier.
As part of this, we want to have files that we intend to keep in sync with
upstream at the same path they have upstream. This includes driver code like
`backend/asmgen.ml{,i}`, but not _true_ backend code, which has more of a clean
break with upstream. With these moves, the files we consider upstreamable will
get conflict markers as usual when merging from upstream, and the files we
consider hard-forked will be untouched (producing modify/delete conflicts, which
we can resolve trivially).

* Handle ocamlformat

Need to move the relevant lines from `backend/.ocamlformat-enable` into
`asmcomp/.ocamlformat-enable` and copy `.ocamlformat` over.