flambda-backend: Continuation lifting (#2295)
Lift continuations (with cli option)

By default, no continuation is lifted, can be configured with a command
line flag. When the continuatin lifting budget is non-zero, we need to
use the advanced meet, because there are known issue that occurs when
using the regular meet together with continuation lifting.

About the fixed test:
Adding a Sys.opaque_identity makes the lifetime of the result
of `quick_stat` does not depend on whether flambda2 lifts some
continuation or not (in turn, this difference of lifetime induces
a difference in the tests results, since it counts the number of
allocations in the major heap).