Typing for parameterised libraries (#1726)
Typechecking of references to parameterised libraries, with or without arguments. Currently such references can't actually be compiled to working code because we don't have the `-instantiate` pass, but this contains all the typechecking logic to make sure that, for instance, a module `A` can't refer to a parameterised module `B` unless `A` takes all the parameters that `B` does (the subset rule) or the reference includes an argument for each missing parameter.