Add partial rebuilding option to ART + some associated refactoring
Instead of encoding a simple boolean, the [Are_rebuilding_terms.t],
now also encodes whether we are rebuilding everything (as is almost
always the case), or whether we are only rebuilding a small part of
the unit (as will be the case for match-in-match/continuation
specialization).

With this changes comes a small refactoring, removing the boolean
conversions to/from [Are_rebuilding_terms] which would be misleading
with the thirs option present, and renaming some functions for better
clarity and consistency.

Finally, the [ART] in [uenv] is only there for printing, could we find
a way to not need it ?