Add a .git-blame-ignore-rev file
`git blame` can be configured to ignore certain commits, such as
formatting commits. This patch adds a `.git-blame-ignore-revs` file (using
a common name) with a few reformatting commits that can be used with
`git blame --ignore-revs-file .git-blame-ignore-revs` or configured as a
default ignore file with:

```
$ git config blame.ignoreRevsFile .git-blame-ignore-revs
```

Hopefully we should rarely need to add new commits to this file (except
on .ocamlformat changes) thanks to better formatting discipline, but
331c1673 (Reformat middle_end/flambda2/) is making `git blame`
effectively useless for archeological tasks. I also heuristically added
a few other commits with "Reformat" in the name, but I have not seen
them actually show up as problematic.