Continuation specialization
Disable specialization in some tests

Interestingly, the blind specialization currently done (since we do not
yet have a heuristic) actually removes the errors on some tests, which
makes the tests fail. We therefore disable specialization on theses
tests and we'll see once we have a proper heuristic if we change/promote
the tests.

Promote test

fix rebase

Prevent continuation specialization when the handler contains a set of
closures