fix(flambda2-types): Prevent duplicate equations in add_equation
The new meet is cautious about keeping track of whether types are
refined during the meet, and uses that information to avoid duplicating
equations.

However, this check is only done when performing a meet between an alias
and a non-alias type, but not when equations are added through
`Typing_env.add_XXX` functions. In particular, equations added when
joining environments (through `add_env_extension_from_level`) are not
deduplicated, which can (theoretically) cause the same equation to
appear at multiple levels.

This patch exposes the `meet_return_value` type to the typing env and
uses it to perform duplication directly in `add_equation`. Since we now
deduplicate in `add_equation`, we don't need to do it during the meet,
which also eliminates a double meet.