Add `-instantiate` (#1905)
This is the last piece of compiler support for parameterised libraries: the ability to take a parameterised module and a set of arguments and produce an instance module. The `-instantiate` option is similar to `-pack` in that it specifies a special compilation mode that takes in .cmo or .cmx files and produces a new one from them. It takes the module to instantiate and the modules to provide as instance arguments.

One bit of awkwardness is that the `-o` option is required, but (up to directory) there's no actual choice: the output module must have the correct form to be usable from other modules so that .cmx lookup can function. This would be somewhat awkward to fix, and it's not worth doing so at the moment since we don't intend for humans to be running `-instantiate` by hand.