Add attributes to (unsafely) skip jkind check  (#3385)
* Add attributes to (unsafely) skip jkind check

Add a pair of attributes, [@@unsafe_allow_any_kind_in_intf] and
[@@unsafe_allow_any_kind_in_impl], which if set on both the impl and the intf
respectively, skip checking the jkind of the type in a signature against the
jkind of the type in a struct entirely. This is a more-selective version of the
`--allow-illegal-crossing` flag, and likely eventually subsumes it.

Signed-off-by: Aspen Smith <aspsmith@janestreet.com>

* Emit a warning when unsafe_allow_any_kind is added unnecessarily

Note that this is /only/ done if the attribute is set in both signatures but not
used - also this is a little over-sensitive (sadly) since this is done during
sigature inclusion too. A new test covers the over-sensitivity.

Signed-off-by: Aspen Smith <aspsmith@janestreet.com>

---------

Signed-off-by: Aspen Smith <aspsmith@janestreet.com>