Hackily sort withs on a kind in outcometree
we need to sort types in a [with] because the order comes from iterating a
TypeHash, which is nondeterministic (specifically, the order is different if we
disable stack allocation) and brittle. Unfortunately, we lack a good comparison
for types, so as a hacky workaround here we format to a string (actually, a
string list for easier line breaking when pretty printing) and sort
lexicographically before finally printing.