Make patricia trees big-endian
This patch switches up the implementation of the `Patricia_tree` module
from little-endian to big-endian, with the main motivation to be able to
implement in-order traversal.

The `caml_int_clz_tagged_to_untagged` and `caml_int_tagged_to_tagged` C
stubs are recognized and replaced with the `clz` instruction when
compiling with flambda2, so they are only used in the boot compiler.