Track unsafe_mode_crossing in type_kind
Add a field to (the relevant variants of) type_kind, which tracks the (extended)
modal upper bounds of a kind introduced by the usage of the
[@@unsafe_allow_any_mode_crossing] attribute. This allows us to check this in eg
the inclusion check, and forbid extending the upper bounds of a type /only/ by
constraining it to a signature. Also, this moves more towards a world where
`decl.type_jkind` is "morally" just a memo of the type's kind, where everything
about the kind /could/ ostensibly be computed by looking at the type_kind.

A few extra test cases checking that we cant do this have also been added to
`allow_any.ml`