Track quality of jkinds
Sometimes, inferred jkinds for a type are not exact, but instead *upper bounds*;
later, we might learn more about a type that causes us to lower that jkind
bound. For example, we might substitute a type variable, or perform `with`
substitution on an abstract type in a module type. It's important to know
whether this might happen, so that we can avoid normalizing away types with
jkinds that might get "better" later from with-bounds during jkind
normalization.

This commit adds a new "jkind_quality" type, and a "quality" field to jkind,
that tracks whether an inferred jkind is "best" or "not best", and sets this to
the correct value everywhere jkinds are constructed. Nothing actually /reads/
this field yet, but this will be used during normalization in a subsequent
commit.