Normalize and deduplicate with-bounds in jkinds
This commit contains (somewhat unfortunately) two distinct, but tightly-related
changes:

1. The internals of with-bounds have been changed from a list to a map (from
   types to sets of axes), with a "best-effort" semantic, structural comparison
   on type_exprs that falls back to pointer comparison on types which are "hard"
   to compare semantically
2. Full, recursive normalization of with-bounds on kinds is implemented,
   respecting quality.

Unfortunately, due to some issues I ran into with stack overflows when compiling
large mutually recursive type declarations (such as the ASTs in the compiler
itself during bootstrap), these ended up as the same commit which it's somewhat
challenging to tease apart. Sorry!

With-bound deduplication
------------------------

For the sake of efficiency, with-bounds have been changed to be a (maybe
non-empty) Map of type_expr, rather than a (maybe non-empty) list. This both
deduplicates them (mostly), which has anywhere from a reasonable to a pretty
significant impact on performance of things that need to deal with
with-bounds (especially for deeply + widely mutually recursive type
declarations, like ASTs), and also makes concatenation faster.

Comparing types semantically (eg, in a way that's compatible with
Ctype.type_equal) is /hard/, though - especially in the context of things like
polymorphic variants and objects. We might someday write an algorithm to do so,
but today this feels like a bit of a monstrous undertaking, and we want to get
something out and working. So instead, we do a hybrid approach: for all the
types which /aren't/ hard to compare (type constructors, type variables, tuples,
etc. - and trees of such) we compare structurally using a recursive
lexicographic tree comparison. For all the remaining "hard" types, we compare
them using pointer equality - but make all "hard" types compare strictly less
than the "easy" ones. This gives us some false positives - we might compare two
types as inequal which are actually semantically equal - but no false negatives,
and also gets us a lot further towards the world we want for the most commonly
used types.

Unfortunately, actually implementing this is a little tricky, because the Map in
stdlib is a functor and the concrete type is not actually exposed, and
with-bounds are mutually recursive (through jkind annotations on type vars) with
type_exprs. We have, as I see it, three options:

1. Two mutually recursive modules, one containing type_expr and one containing
   the functor invocation for the map of types in with_bounds.
2. Define the map as actually just a map of Obj.t, with the comparator behind a
   ref that's map after the recursive type declaration for type_expr and
   with_bounds, and then define another module providing a safe interface
3. Define our own Map type that exposes a concrete type and accepts a
   comparator.

Option 1 is probably the nicest (and likely the only one to be accepted
upstream!) but also involves a *lot* of changes and churn to types.ml, and seems
like it'd make it really hard to pull in changes from upstream in the meantime
due to this churn. Option 3 is also appealing, but a lot of work and finnicky to
get right. So, this commit implements option 2 (with some regret about the Obj)
since it felt like the best point in the tradeoff space

With-bound normalization
------------------------

This commit also implements normalization of with-bounds, adapting and replacing
the existing `extend_bound` function to operate holistically on the whole kind
rather than per-bound, and (optionally!) respect the best-ness of the kinds it
looks up for types in with bounds. This is more or less obvious (I hope!) with a
couple of notable caveats:

1. Different callers of normalize want different behavior when not-best kinds
   are encountered; some want to skip them and leave them in the with-bounds,
   and some want to omit them from the with-bounds but round up the bounds along
   the axes that type matters for to the max. Currently, this behavior is
   exposed via a `~require_best:bool` parameter, which is gross (booleans are
   sad) and really easy to get wrong. It would be preferable to have this
   parameter be a GADT with a phantom `l * r` parameter which controlled whether
   we're expected to return an l- or an lr-jkind, but doing this
   correctly (for... somewhat nuanced reasons) requires correct subsumption, so
   that's *currently* deferred until a follow-up pull request that also adds
   subsumption.

2. For reasons I don't yet entirely understand, I had to change the behavior
   with respect to fuel reuse when normalizing multiple fields of the same type,
   or else I got a stack overflow error when compiling `parsetree.mli` during
   bootstrapping. This is a little sad, but I spent multiple days trying to find
   a better approach and failed, and I'd like to push this forward and revisit
   the fuel usage within normalize later.